+++ mobile/android/fenix/app/src/main/java/org/mozilla/fenix/settings/doh/root/DohSettingsScreen.kt
@@ -72,6 +72,7 @@ import mozilla.components.ui.icons.R as iconsR
  * @param onDefaultInfoClicked Invoked when the user accesses info about Default DoH level.
  * @param onIncreasedInfoClicked Invoked when the user accesses info about Increased DoH level.
  * @param onMaxInfoClicked Invoked when the user accesses info about Max DoH level.
+ * @param onUltraInfoClicked Invoked when the user accesses info about Ultra DoH level.
  */
 @Composable
 internal fun DohSettingsScreen(
@@ -85,6 +86,7 @@ internal fun DohSettingsScreen(
     onDefaultInfoClicked: () -> Unit = {},
     onIncreasedInfoClicked: () -> Unit = {},
     onMaxInfoClicked: () -> Unit = {},
+    onUltraInfoClicked: () -> Unit = {},
 ) {
     Surface {
         Column(
@@ -105,6 +107,7 @@ internal fun DohSettingsScreen(
                 onDefaultInfoClicked = onDefaultInfoClicked,
                 onIncreasedInfoClicked = onIncreasedInfoClicked,
                 onMaxInfoClicked = onMaxInfoClicked,
+                onUltraInfoClicked = onUltraInfoClicked,
             )
 
             HorizontalDivider(
@@ -225,6 +228,7 @@ private fun DohSelection(
     onDefaultInfoClicked: () -> Unit,
     onIncreasedInfoClicked: () -> Unit,
     onMaxInfoClicked: () -> Unit,
+    onUltraInfoClicked: () -> Unit,
 ) {
     state.allProtectionLevels.forEach { protectionLevel ->
         when (protectionLevel) {
@@ -310,6 +314,25 @@ private fun DohSelection(
                     onDohOptionSelected(
                         protectionLevel,
                         state.selectedProvider ?: state.providers.first(),
+                    )
+                },
+            )
+
+            is ProtectionLevel.Ultra -> DohProtectionLevel(
+                modifier = Modifier.fillMaxWidth(),
+                selected = protectionLevel == state.selectedProtectionLevel,
+                label = stringResource(R.string.preference_doh_ultra_protection),
+                summary = stringResource(
+                    R.string.preference_doh_ultra_protection_summary,
+                    stringResource(id = R.string.app_name),
+                ),
+                showInfoIcon = true,
+                provider = null,
+                onInfoClick = onUltraInfoClicked,
+                onClick = {
+                    onDohOptionSelected(
+                        protectionLevel,
+                        null,
                     )
                 },
             )
@@ -547,6 +570,7 @@ private fun DohScreenDefaultProviderPreview(
                     ProtectionLevel.Default,
                     ProtectionLevel.Increased,
                     ProtectionLevel.Max,
+                    ProtectionLevel.Ultra,
                     ProtectionLevel.Off,
                 ),
                 selectedProtectionLevel = ProtectionLevel.Increased,
@@ -574,6 +598,7 @@ private fun DohScreenCustomProviderPreview(
                     ProtectionLevel.Default,
                     ProtectionLevel.Increased,
                     ProtectionLevel.Max,
+                    ProtectionLevel.Ultra,
                     ProtectionLevel.Off,
                 ),
                 selectedProtectionLevel = ProtectionLevel.Increased,